requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
openai==1.50.0
python-dotenv==1.0.1
